requirements.txt
torch>=2.0
torchvision>=0.15
pytorch-lightning>=2.0
tensorboard
numpy
pandas
pyyaml
matplotlib
scikit-learn
pillow
tqdm
